aiservice/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
dashscope==1.14.1
dataclasses-json==0.6.7
DBUtils==3.0.3
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.109.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-checkpoint-postgres==2.0.25
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
openai==2.16.0
orjson==3.11.5
ormsgpack==1.11.0
packaging==23.2
pandas==2.2.0
propcache==0.4.1
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.35.1
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.27.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
